requirements.txt
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
idna==3.4
kafka-python==2.0.2
pydantic==1.10.4
pydantic_core==2.6.1
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
typing_extensions==4.7.1
unicode==2.9
Unidecode==1.3.6
urllib3==2.0.4
